backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
bcrypt==4.0.1
# bzip2==1.0.8
# ca-certificates==2025.1.31
certifi==2025.8.3
# cffi==1.17.1
click==8.2.1
cryptography==46.0.2
# cyrus-sasl==2.1.28
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.0
# expat==2.7.1
fastapi==0.118.0
fastapi-cli==0.0.13
gmpy2==2.2.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
jinja2==3.1.6
# libpq==18.0
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
passlib[bcrypt]==1.7.4
pip==25.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic-core==2.33.2
pyjwt>=2.8.0
pygments==2.19.2
python-dotenv==1.1.1
python-multipart==0.0.20
pyyaml==6.0.3
rich==14.1.0
rich-toolkit==0.15.1
rsa==4.9.1
setuptools==78.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
typer==0.19.2
typing-extensions==4.15.0
typing-inspection==0.4.2
# tzdata==2025b
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
wheel==0.45.1
# xz==5.6.4
# yaml==0.2.5
# zlib==1.2.13
google-generativeai>=0.8.0